Support gather addons
The Addon interface allow adding addons for handling specific resource
types. Addon.Gather() is invoked for every instance of the resource in
the cluster. The logs gatherer is integrated now as an addon.

Addon support allows extending the gather tool to collect more info such
as rook-ceph logs and commands, and integrating with 3rd party tools
such as `subctl gather`.

Because the addons are bound at runtime, we can support running without
any addon or with specific addon.