Disable the useless deprecated warnings
Since we gather all resources we gather also deprecated ones. However
there is no way to skip them without getting the resource from the
server, and the client writes a warning to stderr behind our back:

    W0505 18:54:13.078724  480070 warnings.go:70] kubevirt.io/v1
    VirtualMachineInstancePresets is now deprecated and will be removed
    in v2.

We can disable the warnings to avoid this careless logging. Hopefully it
will not hide meaningful warnings. We can consider enable the warnings
later for development later.