Use zap for logging
The logger is created in the root command, and cloned with Named() for
every cluster. Addons clone it again with the addon name, so we get:

   gather
   gather.dr1
   gather.dr2.logs

Now that we have proper log levels, we have useful INFO level logs for
the command. All other logs converted to DEBUG level.

More work is needed for handling errors - all non-fatal errors should be
logged as warning instead of failing the entire operation.