Gather rook logs from all nodes using agent pod
Gathering from ceph-mgr pod is incorrect:
- on a big cluster, when ceph pods may be running on different nodes, so
  gathering logs form one pod may not gather all logs
- tar is not available in ceph OpenShift images, so we don't gather any
  logs.

Fix both issues using a tiny gather agent pod, created on every node
that rook-ceph pod are running on. We created the temporary pod in the
default namespace, mounting the /data/rook directory, copy the logs to
gather/addons/rook-ceph/log/{node}/ and delete the agent pod.

The agent pod are created in the default namespace. It would be better
to create them in a temporary random namespace to avoid collisions with
real pods.