Fix gathering on OpenShift
On the node the log are stored in different locations:

- OpenShift:  /var/lib/rook/openshift-storage/log
- upstream:   /var/lib/rook/rook-ceph/log
- minikube:   /data/rook/rook-ceph/log

It happens to match the namespace, hopefully this will continue to work.
Another way to find this is to look at one of the ceph pods and find the
volume mounted at /var/lib/ceph/log. Using the namespace is simpler so
lets keep using it.