Fix getting previous logs
Previous log is available only if there is a terminated pod. We can tell
by looking in the .status.containerStatues. If the status has:

    lastState:
      terminated:
        containerID: ...

We can get a previous log for the container. We extract this info when
listing containers, and queue gathering for previous log only when the
log is available.