Fix error handling when gathering logs
No error is fatal in this case. If something fails, we log a warning and
continue.

When streaming logs data, the request can succeed only if we ask a
previous log and the log exist, or if the container is running. We get a
BadRequest error even when the request is fine, and have no way to
detect the root cause. Therefore we log the error as debug message to
avoid snapping the logs with false warnings