Make log required for RemoteDirectory
Make fields private and add a NewRemoteDirectory function. Instead of
generic strings, use Pod and Options, so it is easier to create from
addons.

With this we can remove the logging wrapper and use the log directly,
which make it easier to manage logging calls, and allow using warnings
without adding more wrappers.