Validate namespace before gathering
If a namespace was specified and it does not exist on the cluster, there
is no need to list api resources and then try to list all resources in
the missing namespaces. On OpenShift ODF cluster this will try to list
170 api resources in a non-existing namespace which is pointless and
slow.

A typical use case is gathering an DR protected application namespace.
In some cases we have the namespace only one managed clusters, on both
managed clusters, or on all clusters. We want to help the user gather
all the available info about the application regardless of the
implementation details.

If the user specify the wrong namespace name, nothing will be gathered
very quickly without any warning. Maybe we can detect this and log a
warning, but this seems too much effort for unlikely use case, so maybe
later.