Report number of gathered resources
This gives a nice overview of the amount of work done, and allows
warning when nothing was gathered from any cluster, for example when
using the wrong namespace. Maintaining a resource count will also allow
showing progress for the operation later.

Example run on OpenShift Regional DR environment:

    $ kubectl gather -d gather --contexts perf4,perf5,perf8
    ...
    2024-05-11T02:45:42.535+0300	INFO	gather	Gathered 11527 resources from cluster "perf4" in 83.447 seconds
    2024-05-11T02:46:31.747+0300	INFO	gather	Gathered 11074 resources from cluster "perf5" in 132.659 seconds
    2024-05-11T02:46:53.134+0300	INFO	gather	Gathered 11060 resources from cluster "perf8" in 154.046 seconds
    2024-05-11T02:46:53.134+0300	INFO	gather	Gathered 33661 resources from 3 clusters in 154.050 seconds

If namespace was specified and nothing was gathered from any cluster, we
log a warning:

    $ kubectl gather -d gather --contexts dr1,dr2,hub -n no-such-namespace
    ...
    2024-05-11T03:29:31.792+0300	WARN	gather	No resource gathered from namespace "no-such-namespace"
    2024-05-11T03:29:31.792+0300	INFO	gather	Gathered 0 resources from 3 clusters in 0.009 seconds

We don't fail as this may be a valid use case.