Handle ADDED event
When starting watching we gat a watch.Added  event immediately, and then
watch.Modified event on changes. In practice it the pod is never in
running in the Added event, but theoretically it can be if there was
some delay in creating the watcher. Now we check the phase in both
events.