Improve gather directory help and format
Previously the default directory showed the special template value:

    -d, --directory string    directory for storing gathered data (default "gather-20060102150405")

Which is never the actual value. Now we use zero value as default, and
specify a default in the actual flag help:

    -d, --directory string    directory for storing gathered data (default "gather.{timestamp}")

The gather directory is generated after the command is run if the value
is empty. To help users find the right directory, log the generated
directory name:

    $ kubectl gather -n foo
    2024-05-24T20:14:14.442+0300	INFO	gather	Storing data in "gather.20240524201414"

The default name use now gather.timestamp instead of gather-timestamp to
make the name different from user selected name, which are likely to use
xxx-yyy-zzz format (like most names in k8s).