Simpler and more robust namespace validation
Previously we replaced opts.Namespaces during validation. This is
fragile since empty namespaces means collect all namespaces. It worked
since we terminate when there is no namespace to gather, but fragile.

Now instead of modifying the options, we return the list of existing
namespaces. This is more robust and simpler.