Limit memory and cpu usage with paging
When listing resources on big clusters we may get large number of items
in a single request. This is bad both for the server producing the huge
JSON request, and for us parsing and processing huge JSON payload. Use
paging API[1] to limit number of items returned for single list request.

[1] https://kubernetes.io/docs/reference/using-api/api-concepts/#retrieving-large-results-sets-in-chunks