Streamline resourceInfo
Instead of keeping pointer to GropuVersin and APIResource, embed a
GroupVersionResource and add Namespaced boolean.

With this we have all the info we need for dumping and listing resources
in the resource info struct, and we can minimize allocations and improve
locality. Code is also little bit simpler since it can access the
embedded GroupVersionResource directly.