Eliminate addon name duplication
Keep the addon name in a constant in both addons to have more uniform
code and remove duplication. This way the same name is used in logs,
enabling addons, and command names. The addons code is not more
uniform so it is easier to learn how to create a new addon.