Introduce AddonBackend interface
Instead of passing 5 arguments to AddonFunc, pass an AddonBackend
interface providing functions to get the arguments.

The AddonBackend interface extends the Queuer interface so addons can
embed the backend and use it to get options, output directory, and queue
work. This simplifies the addon types, cleans up the code and make it
more uniform.

The motivation is:
- make it easier to create new addons using simpler interfaces
- make addons code easier to understand since existing addons are the
  example for creating new addons.
- support concurrent gathering eliminating duplicate gathering for the
  same resource
- make it possible to move addons to a new package by depending on few
  public gather types and interfaces