Fix error handling when terminating remote tar
When we terminate the remote tar command after local tar command failed
to start, check and log the unlikely but possible error to kill the
remote tar command.

If we could not kill the command, don't try to wait for it since it will
block forever anyway.

Explicitly ignore the return value of Wait() since the command was
terminate by signal.