Use make during release
The build command is duplicated in the release workflow, Makefile, and
Containerfile. The command is simple now so this is fine, but we want to
make inject build info into the executable. Simplify by using make in
the release workflow.

This also makes debugging release issues easier since it uses the same
command we can run locally.