New gather pipeline
Eliminate the deadlocks and uncontrolled memory usage by switching to
unbuffered channels. When queuing work, the function returns when one of
the workers picked up the work. If all workers are busy, the caller is
blocked until a worker is available.

Previously we had one work queue, and anyone could queue work on the
queue for later processing without any limit. This caused deadlocks
when al workers are trying to queue work and the queue becomes full.

The new pipeline splits the work to 3 steps:

    prepare -> gather -> inspect

Pipeline steps:

1. prepare: look up namespaces and api resources, and queue work on the
   gather work queue.
2. gather: run work functions queued by the prepare step, and queue work
   on the inspect work queue
3. inspect: run work functions queued by the gather step.

Since every step queues work on the next step queue, workers cannot
deadlock.

Waiting for completion is simpler and more robust now:

- When the prepare step completes, we know that no more work can be
  queued in the gather work queue, so we close it.
- When the gather work queue completes, we know that no more work can be
  queued in the inspect work queue so we close it.
- When the inspect work queue completes, we know that all work was
  finished and we can return.

The limitation with the new pipeline is that you cannot queue more work
from a queued work function on the same queue. In every step you must
get all the data you need for queueing work on the next step. For
example when inspecting pods, we must get all the containers from the
pod, and queue gather log function on the inspect work queue. This
limitation is not an issue for the current code.

Previously we had one work queue with 6 workers. Now we use 2 work
queues with 6 workers in each. Testing shows that number of workers does
not affect memory usage significantly, but it decrease the time to
gather.

Comparing to previous version when gathering entire ODF managed cluster,
collecting 3.1 GiB of data. The new version is 1.3 times faster but use
1.13 times more memory.

| version | maximum RSS | time      |
|---------|-------------|-----------|
| before  |   130.9 MiB |   228.6 s |
| after   |   148.0 MiB |   173.3 s |

Another advantage of the new pipeline is that work queues are closed
when Gather() returns. Previously we leaked the goroutines until the
process terminates. This is not an issue for kubectl-gather command, but
it is an issue when using kubectl-gather as a library, for example
gathering data after every test step in ramenctl.