build: Make golang container version configurable
Currently we need to update the version manually each time go is
upgraded. Replace the hard code version with a build variable.

We pass the go version from the go.mod file to the container for
selecting the build container. This matches the go version used in the
github runner.