gather: Introduce OutputReader
OutputReader makes it easy to list and read gathered resources from the
output directory.

Like OutputDirectory is works on a single gather directory. To read read
multiple clusters data create multiple readers. It may more useful to
support multiple directories, but inconsistent with current gather API.

The reader provide the following functions:

- ListResources(namespace, resource) - returns list of resource names.
  This is required when the resource name contains random part.
- ReadResource(namespace, resource, name) - read resource data

Empty namespace access cluster scope resources.

I considered adding OpenResource() but this add lot of boilerplate to
read the data and closing the file, and resources are small so it is not
helpful.